CI/release: Add back wildcard match for tags
While I've always thought that the wildcard was _not_ required if any
tag is accepted, it does however make sure that GitHub does _not_ run
on any branch push.  Right now it does so, because it probably sees
an empty `tags:` object and considers that to be the same as an empty
`push:` object.

As can be seen on the latest merge to `master`, the release workflow
tried to run (but failed because there was no accompanying tag).